+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Monitor representations and video modes.
 - Cleaning functions that clean the built up memory usage of the engine.
 - `is_initialized` function for object.
+- New networking system feature to communicate between clients.
 
 ### Changed
 
@@ -31,6 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated Rapier
 - Log crate macros instead of println are getting used for the vulkan validation layer feature.
 - Updated winit, therefore updated `keycode` to `key`and added most of the key parts to the `KeyboardEvent` struct.
+- `Game` functions are now all async, except for `exit`
 
 ### Fixed